requirements.txt
chardet==5.2.0

pandas==2.3.3
mysql-connector-python==9.5.0
selenium==4.38.0
beautifulsoup4==4.14.2